requirements-local.txt
# Local development requirements with compatible versions for Groq API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13  # Compatible version with OpenAI library
openai==1.3.5  # Locked version to prevent breaking changes
aiokafka==0.8.1
asyncpg==0.29.0
twilio==8.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
mcp==1.0.0
anthropic==0.21.3
python-multipart==0.0.6
cryptography==41.0.8
numpy==1.24.3
psycopg2-binary==2.9.9
pgvector==0.2.4
websockets==12.0
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
locust==2.20.0